fix negative x & y

+++ app/src/main/kotlin/silver/ore/core/ClusterId.kt
@@ -1,6 +1,6 @@
 package silver.ore.core
 
-data class ClusterId(var x: Int, var y: Int) {
+data class ClusterId(var x: Long, var y: Long) {
     fun getNeighbourhood(): Array<ClusterId> {
         return arrayOf(ClusterId(x, y-1),
                        ClusterId(x+1, y-1),
