+++ app/src/main/kotlin/silver/ore/core/Map.kt
@@ -3,14 +3,14 @@ package silver.ore.core
 import kotlin.random.Random
 
 class Map(val random: Random) {
-    private val width = 16
-    private val height = 16
+    private val width = 16L
+    private val height = 16L
     private val tiles = HashMap<ClusterId, Tile>()
     val humanTownClusterId: ClusterId
 
     init {
-        val x = random.nextInt(0, width)
-        val y = random.nextInt(0, height)
+        val x = random.nextLong(0, width)
+        val y = random.nextLong(0, height)
         humanTownClusterId = ClusterId(x, y)
         tiles[humanTownClusterId] = Tile(humanTownClusterId, Tile.TYPE.TOWN)
     }
